deal-scanner/requirements.txt
# Core Dependencies
python-dotenv==1.0.0
requests==2.32.4
beautifulsoup4==4.12.2
lxml==4.9.3

# Web Scraping
selenium==4.15.2
webdriver-manager==4.0.1
playwright==1.40.0

# Agent Framework
langgraph==0.0.26
langchain==0.2.5
langchain-openai==0.0.2
langchain-community==0.3.27

# Database
sqlalchemy==2.0.23

# Scheduling
schedule==1.2.0
APScheduler==3.10.4

# RSS Feed Parsing
feedparser==6.0.10

# Telegram Bot
python-telegram-bot==20.7

# HTTP Clients & Rate Limiting
httpx==0.25.2
aiohttp==3.12.14
tenacity==8.2.3

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Logging
loguru==0.7.2

# API Clients
openai==1.6.1

# Utilities
pydantic==2.5.2
python-dateutil==2.8.2
pytz==2023.3
fake-useragent==1.4.0
